+++ src/utility/cueinterpreter.cpp
@@ -251,7 +251,7 @@ namespace PixelMaestroStudio {
 		switch((MaestroCueHandler::Action)cue[(uint8_t)MaestroCueHandler::Byte::ActionByte]) {
 			case MaestroCueHandler::Action::SetBrightness:
 				result->append(": ");
-				result->append(cue[(uint8_t)MaestroCueHandler::Byte::OptionsByte]);
+				result->append(QString::number(cue[(uint8_t)MaestroCueHandler::Byte::OptionsByte]));
 				break;
 			case MaestroCueHandler::Action::SetShow:
 				break;