+++ src/widget/devicecontrolwidget.h
@@ -75,7 +75,7 @@ namespace PixelMaestroStudio {
 			void disconnect_serial_device(int index);
 			void populate_serial_devices();
 			void set_device_controls_enabled(bool enabled);
-			void write_to_device(SerialDeviceController* device, const char* out, int size, bool progress = false);
+			void write_to_device(SerialDeviceController* device, const char* out, const int size, bool progress = false);
 	};
 }